Suppress Clang warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@103264 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netschedule/ns_precise_time.hpp
@@ -72,7 +72,7 @@ class CNSPreciseTime : public timespec
           tv_nsec = 0; }
         CNSPreciseTime(double  time)
         { tv_sec = int(time);
-          tv_nsec = int( (time - tv_sec) * kNSecsPerSecond ); }
+          tv_nsec = int( (time - tv_sec) * (long)(kNSecsPerSecond) ); }
         CNSPreciseTime(unsigned int sec, unsigned int  nsec)
         { tv_sec = sec;
           tv_nsec = nsec; }
